style(trusty-mpm): fmt + move #1809/#3034 tombstone-filter tests under line cap (#3044)
CI hygiene-gate fixes for #3044:

- `cargo fmt --all` — reformats indentation left over from the earlier
  conflict-marker resolution (plain string edits, not `Edit` tool passes,
  in a few struct literals under `tui/project_ctl`).
- The #3044 reconciliation's added test coverage pushed
  `tests_behavior_c_tests.rs` to 1577 SLOC, over the 1500-SLOC test-file
  cap (`scripts/check_line_cap.sh`). Moved the `#1809`/`#3034`/`#3044`
  decommissioned/deleted-tombstone picker-filter tests
  (`is_live_session_state_*`, `picker_filter_*`) to
  `tests_behavior_e_tests.rs`, which had headroom under the same
  `tests_behavior_a/b/c/d/e` split convention this repo already uses.
  No test or logic was deleted — same 7 tests, same assertions, new home.

`scripts/check_line_cap.sh`: 7 allowlisted, 0 violations — OK.
`cargo fmt --all --check`: clean.
`cargo build -p trusty-mpm` / `cargo clippy -p trusty-mpm --all-targets -- -D warnings`: clean.
