+++ crates/trusty-mpm/src/tui/project_ctl/poll/rows.rs
@@ -215,9 +215,9 @@ mod tests {
             injection_status: None,
             unresumable: false,
             stale_assets: false,
-attached: false,
-slot: 0,
-deleted: false,
+            attached: false,
+            slot: 0,
+            deleted: false,
         }
     }
